Rating summary added
Added rating summary function that displays a summary on a specific element with id 'data-surrounding-short-element-id'. If it's present then it should display the summary rating there. This script will also add a 'json-ld' script to each site where it could find the surrounding element. This is used for Google Rich Snippets.